# public/openapi-everything.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AB Brain - Complete API",
...
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://ab-brain.up.railway.app/api"
    }
  ],
  "paths": {
...
    "/daily-plans": {
      "get": {
        "operationId": "listDailyPlans",
        "summary": "List daily plans with date range filters",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Exact date to match"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date for range"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date for range"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "planned",
                "completed",
                "partial",
                "missed",
                "rest",
                "amended"
              ]
            }
          },
          {
            "name": "training_plan_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of daily plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DailyPlan"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createDailyPlan",
        "summary": "Create a single daily plan with workout, nutrition, and recovery targets",
        "description": "Creates one daily plan. plan_date is required and must be unique. Use POST /daily-plans/week to create 7 plans at once. Targets set here override gamification_settings defaults for ring calculations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyPlanInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created daily plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPlan"
                }
              }
            }
          },
          "409": {
            "description": "A plan already exists for this date. Use PUT to amend it."
          }
        }
      }
    },
    "/daily-plans/week": {
      "post": {
        "operationId": "createWeeklyPlan",
        "summary": "Create 7 daily plans at once for a week",
        "description": "Accepts start_date and an array of 1-7 daily plan objects. Each is assigned consecutive dates starting from start_date. If a plan already exists for a date, it is updated (upsert). Link to a training_plan_id to associate with a mesocycle/block.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "start_date",
                  "days"
                ],
                "properties": {
                  "start_date": {
                    "type": "string",
                    "format": "date",
                    "description": "First day of the week (usually Monday)"
                  },
                  "training_plan_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Optional link to parent training plan"
                  },
                  "days": {
                    "type": "array",
                    "description": "Array of 1-7 daily plan objects. Each gets consecutive dates starting from start_date.",
                    "items": {
                      "$ref": "#/components/schemas/DailyPlanInput"
                    },
                    "minItems": 1,
                    "maxItems": 7
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created": {
                      "type": "integer"
                    },
                    "plans": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DailyPlan"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/daily-plans/by-date/{date}": {
      "get": {
        "operationId": "getDailyPlanByDate",
        "summary": "Get daily plan + actual data + plan-vs-actual comparison + ring progress for a date",
        "description": "The key endpoint for the daily fitness workflow. Returns the plan, all actual logged data (workouts, meals, nutrition context, body metrics), a structured comparison, and ring progress calculations.",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {
            "description": "Plan + actual + comparison + ring progress",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "$ref": "#/components/schemas/DailyPlan"
                    },
                    "actual": {
                      "type": "object",
                      "properties": {
                        "workouts": {
                          "type": "array"
                        },
                        "meals": {
                          "type": "array"
                        },
                        "nutrition_context": {
                          "type": "object"
                        },
                        "body_metrics": {
                          "type": "array"
                        }
                      }
                    },
                    "comparison": {
                      "type": "object",
                      "description": "Actual vs target for effort, calories, protein, hydration, sleep"
                    },
                    "ring_progress": {
                      "type": "object",
                      "description": "Train/Fuel/Recover ring percentages and hit/miss details"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/daily-plans/{id}": {
      "put": {
        "operationId": "updateDailyPlan",
        "summary": "Amend/update a daily plan",
        "description": "Use this to adjust today's or future plans based on how things are going. Can change targets, status, notes, etc.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyPlanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPlan"
                }
              }
            }
          }
        }
      }
    },
...
    "/daily-plans/{id}/review": {
      "get": {
        "operationId": "reviewDailyPlan",
        "summary": "Review a daily plan with actual results \u2014 plan vs actual comparison with workouts, meals, nutrition, injuries, and summary stats",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan with actual data and comparison summary"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DailyPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan_date": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "completed",
              "partial",
              "missed",
              "rest",
              "amended"
            ]
          },
          "title": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "workout_type": {
            "type": "string"
          },
          "workout_focus": {
            "type": "string"
          },
          "target_effort": {
            "type": "integer"
          },
          "target_duration_min": {
            "type": "integer"
          },
          "workout_notes": {
            "type": "string"
          },
          "target_calories": {
            "type": "number"
          },
          "target_protein_g": {
            "type": "number"
          },
          "target_carbs_g": {
            "type": "number"
          },
          "target_fat_g": {
            "type": "number"
          },
          "target_hydration_liters": {
            "type": "number"
          },
          "target_sleep_hours": {
            "type": "number"
          },
          "recovery_notes": {
            "type": "string"
          },
          "coaching_notes": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ai_source": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DailyPlanInput": {
        "type": "object",
        "required": [
          "plan_date"
        ],
        "properties": {
          "plan_date": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "completed",
              "partial",
              "missed",
              "rest",
              "amended"
            ],
            "default": "planned"
          },
          "title": {
            "type": "string",
            "description": "Plan title (e.g. 'Friday \u2014 Quality Run')"
          },
          "goal": {
            "type": "string",
            "description": "What this day's plan aims to achieve"
          },
          "workout_type": {
            "type": "string",
            "description": "Type: strength, hill, run, hybrid, recovery, ruck, hiit, etc."
          },
          "workout_focus": {
            "type": "string",
            "description": "Specific focus (e.g. 'Upper Push', 'Pacing Discipline')"
          },
          "target_effort": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "target_duration_min": {
            "type": "integer"
          },
          "workout_notes": {
            "type": "string"
          },
          "target_calories": {
            "type": "number"
          },
          "target_protein_g": {
            "type": "number"
          },
          "target_carbs_g": {
            "type": "number"
          },
          "target_fat_g": {
            "type": "number"
          },
          "target_hydration_liters": {
            "type": "number"
          },
          "target_sleep_hours": {
            "type": "number"
          },
          "recovery_notes": {
            "type": "string"
          },
          "coaching_notes": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ai_source": {
            "type": "string",
            "default": "chatgpt"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}